+++ docs/guides/code_examples/avoid_blocking/default_fingerprint_generator_with_args.py
