+++ docs/guides/code_examples/playwright_crawler_adaptive/init_beautifulsoup.py
@@ -1,0 +1,15 @@
+from crawlee.crawlers import AdaptivePlaywrightCrawler
+
+
+async def main() -> None:
+    crawler = AdaptivePlaywrightCrawler.with_beautifulsoup_static_parser(
+        # Arguments relevant only for PlaywrightCrawler
+        playwright_crawler_specific_kwargs={
+            'headless': False,
+            'browser_type': 'chromium',
+        },
+        # Common arguments relevant to all crawlers
+        max_crawl_depth=5,
+    )
+
+    # ...
