+++ docs/guides/code_examples/scaling_crawlers/max_tasks_per_minute_example.py
@@ -15,3 +15,5 @@ async def main() -> None:
         # Apply the defined concurrency settings to the crawler.
         concurrency_settings=concurrency_settings,
     )
+
+    # ...
