+++ docs/guides/code_examples/scaling_crawlers/min_and_max_concurrency_example.py
@@ -17,3 +17,5 @@ async def main() -> None:
         # Use the configured concurrency settings for the crawler.
         concurrency_settings=concurrency_settings,
     )
+
+    # ...
