+++ docs/guides/code_examples/session_management/sm_http.py
