+++ docs/guides/http_clients.mdx
@@ -7,10 +7,10 @@ description: Crawlee supports multiple HTTP clients when making requests.
 import ApiLink from '@site/src/components/ApiLink';
 import Tabs from '@theme/Tabs';
 import TabItem from '@theme/TabItem';
-import CodeBlock from '@theme/CodeBlock';
+import RunnableCodeBlock from '@site/src/components/RunnableCodeBlock';
 
-import BsCurlImpersonateExample from '!!raw-loader!./code_examples/http_clients/curl_impersonate_example.py';
-import BsHttpxExample from '!!raw-loader!./code_examples/http_clients/httpx_example.py';
+import BsCurlImpersonateExample from '!!raw-loader!roa-loader!./code_examples/http_clients/curl_impersonate_example.py';
+import BsHttpxExample from '!!raw-loader!roa-loader!./code_examples/http_clients/httpx_example.py';
 
 HTTP clients are utilized by the HTTP-based crawlers (e.g. <ApiLink to="class/BeautifulSoupCrawler">`BeautifulSoupCrawler`</ApiLink>) to communicate with web servers. They use external HTTP libraries for communication, rather than a browser. Examples of such libraries include [httpx](https://pypi.org/project/httpx/), [aiohttp](https://pypi.org/project/aiohttp/) or [curl-cffi](https://pypi.org/project/curl-cffi/). After retrieving page content, an HTML parsing library is typically used to facilitate data extraction. Examples of such libraries are [beautifulsoup](https://pypi.org/project/beautifulsoup4/), [parsel](https://pypi.org/project/parsel/), [selectolax](https://pypi.org/project/selectolax/) or [pyquery](https://pypi.org/project/pyquery/). These crawlers are faster than browser-based crawlers but they cannot execute client-side JavaScript.
 
@@ -20,14 +20,14 @@ In Crawlee we currently have two HTTP clients: <ApiLink to="class/HttpxHttpClien
 
 <Tabs>
     <TabItem value="BeautifulSoupHttpxExample" label="BeautifulSoupCrawler with HTTPX">
-        <CodeBlock className="language-python">
+        <RunnableCodeBlock className="language-python" language="python">
             {BsHttpxExample}
-        </CodeBlock>
+        </RunnableCodeBlock>
     </TabItem>
     <TabItem value="BeautifulSoupCurlImpersonateExample" label="BeautifulSoupCrawler with Curl impersonate">
-        <CodeBlock className="language-python">
+        <RunnableCodeBlock className="language-python" language="python">
             {BsCurlImpersonateExample}
-        </CodeBlock>
+        </RunnableCodeBlock>
     </TabItem>
 </Tabs>
