+++ docs/guides/playwright_crawler.mdx
@@ -6,11 +6,13 @@ description: How to use the PlaywrightCrawler and its related components.
 
 import ApiLink from '@site/src/components/ApiLink';
 import CodeBlock from '@theme/CodeBlock';
+import RunnableCodeBlock from '@site/src/components/RunnableCodeBlock';
 
-import MultipleLaunchExample from '!!raw-loader!./code_examples/playwright_crawler/multiple_launch_example.py';
-import BrowserConfigurationExample from '!!raw-loader!./code_examples/playwright_crawler/browser_configuration_example.py';
+import MultipleLaunchExample from '!!raw-loader!roa-loader!./code_examples/playwright_crawler/multiple_launch_example.py';
+import BrowserConfigurationExample from '!!raw-loader!roa-loader!./code_examples/playwright_crawler/browser_configuration_example.py';
+import PreNavigationExample from '!!raw-loader!roa-loader!./code_examples/playwright_crawler/pre_navigation_hook_example.py';
+
 import PluginBrowserConfigExample from '!!raw-loader!./code_examples/playwright_crawler/plugin_browser_configuration_example.py';
-import PreNavigationExample from '!!raw-loader!./code_examples/playwright_crawler/pre_navigation_hook_example.py';
 
 A <ApiLink to="class/PlaywrightCrawler">`PlaywrightCrawler`</ApiLink> is a browser-based crawler. In contrast to HTTP-based crawlers like <ApiLink to="class/ParselCrawler">`ParselCrawler`</ApiLink> or <ApiLink to="class/BeautifulSoupCrawler">`BeautifulSoupCrawler`</ApiLink>, it uses a real browser to render pages and extract data. It is built on top of the [Playwright](https://playwright.dev/python/) browser automation library. While browser-based crawlers are typically slower and less efficient than HTTP-based crawlers, they can handle dynamic, client-side rendered sites that standard HTTP-based crawlers cannot manage.
 
@@ -35,17 +37,17 @@ The <ApiLink to="class/PlaywrightCrawler">`PlaywrightCrawler`</ApiLink> uses oth
 
 The <ApiLink to="class/BrowserPool">`BrowserPool`</ApiLink> allows you to manage multiple browsers. Each browser instance is managed by a separate <ApiLink to="class/PlaywrightBrowserPlugin">`PlaywrightBrowserPlugin`</ApiLink> and can be configured independently. This is useful for scenarios like testing multiple configurations or implementing browser rotation to help avoid blocks or detect different site behaviors.
 
-<CodeBlock className="language-python">
+<RunnableCodeBlock className="language-python" language="python">
     {MultipleLaunchExample}
-</CodeBlock>
+</RunnableCodeBlock>
 
 ## Browser launch and context configuration
 
 The <ApiLink to="class/PlaywrightBrowserPlugin">`PlaywrightBrowserPlugin`</ApiLink> provides access to all relevant Playwright configuration options for both [browser launches](https://playwright.dev/python/docs/api/class-browsertype#browser-type-launch) and [new browser contexts](https://playwright.dev/python/docs/api/class-browser#browser-new-context). You can specify these options in the constructor of <ApiLink to="class/PlaywrightBrowserPlugin">`PlaywrightBrowserPlugin`</ApiLink> or <ApiLink to="class/PlaywrightCrawler">`PlaywrightCrawler`</ApiLink>:
 
-<CodeBlock className="language-python">
+<RunnableCodeBlock className="language-python" language="python">
     {BrowserConfigurationExample}
-</CodeBlock>
+</RunnableCodeBlock>
 
 You can also configure each plugin used by <ApiLink to="class/BrowserPool">`BrowserPool`</ApiLink>:
 
@@ -59,9 +61,9 @@ For an example of how to implement a custom browser plugin, see the [Camoufox ex
 
 In some use cases, you may need to configure the [page](https://playwright.dev/python/docs/api/class-page) before it navigates to the target URL. For instance, you might set navigation timeouts or manipulate other page-level settings. For such cases you can use the <ApiLink to="class/PlaywrightCrawler#pre_navigation_hook">`pre_navigation_hook`</ApiLink> method of the <ApiLink to="class/PlaywrightCrawler">`PlaywrightCrawler`</ApiLink>. This method is called before the page navigates to the target URL and allows you to configure the page instance.
 
-<CodeBlock className="language-python">
+<RunnableCodeBlock className="language-python" language="python">
     {PreNavigationExample}
-</CodeBlock>
+</RunnableCodeBlock>
 
 ## Conclusion
