+++ docs/guides/proxy_management.mdx
@@ -8,16 +8,18 @@ import ApiLink from '@site/src/components/ApiLink';
 import Tabs from '@theme/Tabs';
 import TabItem from '@theme/TabItem';
 import CodeBlock from '@theme/CodeBlock';
+import RunnableCodeBlock from '@site/src/components/RunnableCodeBlock';
 
-import QuickStartExample from '!!raw-loader!./code_examples/proxy_management/quick_start_example.py';
-import IntegrationBsExample from '!!raw-loader!./code_examples/proxy_management/integration_bs_example.py';
-import IntegrationPwExample from '!!raw-loader!./code_examples/proxy_management/integration_pw_example.py';
+import QuickStartExample from '!!raw-loader!roa-loader!./code_examples/proxy_management/quick_start_example.py';
+import IntegrationBsExample from '!!raw-loader!roa-loader!./code_examples/proxy_management/integration_bs_example.py';
+import IntegrationPwExample from '!!raw-loader!roa-loader!./code_examples/proxy_management/integration_pw_example.py';
+import TiersBsExample from '!!raw-loader!roa-loader!./code_examples/proxy_management/tiers_bs_example.py';
+import TiersPwExample from '!!raw-loader!roa-loader!./code_examples/proxy_management/tiers_pw_example.py';
+import InspectionBsExample from '!!raw-loader!roa-loader!./code_examples/proxy_management/inspecting_bs_example.py';
+import InspectionPwExample from '!!raw-loader!roa-loader!./code_examples/proxy_management/inspecting_pw_example.py';
+
 import SessionBsExample from '!!raw-loader!./code_examples/proxy_management/session_bs_example.py';
 import SessionPwExample from '!!raw-loader!./code_examples/proxy_management/session_pw_example.py';
-import InspectionBsExample from '!!raw-loader!./code_examples/proxy_management/inspecting_bs_example.py';
-import InspectionPwExample from '!!raw-loader!./code_examples/proxy_management/inspecting_pw_example.py';
-import TiersBsExample from '!!raw-loader!./code_examples/proxy_management/tiers_bs_example.py';
-import TiersPwExample from '!!raw-loader!./code_examples/proxy_management/tiers_pw_example.py';
 
 [IP address blocking](https://en.wikipedia.org/wiki/IP_address_blocking) is one of the oldest and most effective ways of preventing access to a website. It is therefore paramount for a good web scraping library to provide easy to use but powerful tools which can work around IP blocking. The most powerful weapon in our anti IP blocking arsenal is a [proxy server](https://en.wikipedia.org/wiki/Proxy_server).
 
@@ -29,9 +31,9 @@ With Crawlee we can use our own proxy servers or proxy servers acquired from thi
 
 If you already have proxy URLs of your own, you can start using them immediately in only a few lines of code.
 
-<CodeBlock className="language-python">
+<RunnableCodeBlock className="language-python" language="python">
     {QuickStartExample}
-</CodeBlock>
+</RunnableCodeBlock>
 
 Examples of how to use our proxy URLs with crawlers are shown below in [Crawler integration](#crawler-integration) section.
 
@@ -45,14 +47,14 @@ All our proxy needs are managed by the <ApiLink to="class/ProxyConfiguration">`P
 
 <Tabs>
     <TabItem value="BeautifulSoupCrawler" label="BeautifulSoupCrawler">
-        <CodeBlock className="language-python">
+        <RunnableCodeBlock className="language-python" language="python">
             {IntegrationBsExample}
-        </CodeBlock>
+        </RunnableCodeBlock>
     </TabItem>
     <TabItem value="PlaywrightCrawler" label="PlaywrightCrawler">
-        <CodeBlock className="language-python">
+        <RunnableCodeBlock className="language-python" language="python">
             {IntegrationPwExample}
-        </CodeBlock>
+        </RunnableCodeBlock>
     </TabItem>
 </Tabs>
 
@@ -89,14 +91,14 @@ In an active tier, Crawlee will alternate between proxies in a round-robin fashi
 
 <Tabs>
     <TabItem value="BeautifulSoupCrawler" label="BeautifulSoupCrawler">
-        <CodeBlock className="language-python">
+        <RunnableCodeBlock className="language-python" language="python">
             {TiersBsExample}
-        </CodeBlock>
+        </RunnableCodeBlock>
     </TabItem>
     <TabItem value="PlaywrightCrawler" label="PlaywrightCrawler">
-        <CodeBlock className="language-python">
+        <RunnableCodeBlock className="language-python" language="python">
             {TiersPwExample}
-        </CodeBlock>
+        </RunnableCodeBlock>
     </TabItem>
 </Tabs>
 
@@ -106,13 +108,13 @@ The <ApiLink to="class/BeautifulSoupCrawler">`BeautifulSoupCrawler`</ApiLink> an
 
 <Tabs>
     <TabItem value="BeautifulSoupCrawler" label="BeautifulSoupCrawler">
-        <CodeBlock className="language-python">
+        <RunnableCodeBlock className="language-python" language="python">
             {InspectionBsExample}
-        </CodeBlock>
+        </RunnableCodeBlock>
     </TabItem>
     <TabItem value="PlaywrightCrawler" label="PlaywrightCrawler">
-        <CodeBlock className="language-python">
+        <RunnableCodeBlock className="language-python" language="python">
             {InspectionPwExample}
-        </CodeBlock>
+        </RunnableCodeBlock>
     </TabItem>
 </Tabs>
